pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "e3-django-dynamic-forms"
version = "0.1.4"
description = "A dynamic JSON-based form engine for Django"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Nestor Bracho" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["django", "forms", "dynamic", "json", "schema"]
dependencies = [
    "Django>=4.1",
    "djangorestframework>=3.14",
]

[project.optional-dependencies]
images = ["Pillow"]
dev = [
    "pytest",
    "pytest-django",
]

[project.urls]
Homepage = "https://github.com/e3tools/e3-django-dynamic-forms"
Repository = "https://github.com/e3tools/e3-django-dynamic-forms"
"Bug Tracker" = "https://github.com/e3tools/e3-django-dynamic-forms/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/e3_dynamic_forms"]

[tool.hatch.build.targets.sdist]
include = [
    "src/e3_dynamic_forms",
    "LICENSE",
    "README.md",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.test_app.settings"
pythonpath = ["src", "."]
